requirements.txt
pyrogram
pyromod
pytz
pillow
tgcrypto